requirements.txt
# Core dependencies
crewai>=0.1.31
openai>=1.3.0
python-dotenv>=0.19.0

# Data handling and analysis
pandas>=2.0.0
pyyaml>=6.0.0
pydantic>=2.0.0

# Jupyter/IPython (for display functionality)
ipython>=8.0.0
jupyter>=1.0.0

# Type hints
typing-extensions>=4.0.0

# Optional
black>=23.0.0  # Code formatting
pylint>=2.17.0  # Code linting
pytest>=7.0.0  # Testing